revert exception handling
