+++ src/Serilog.Sinks.OpenTelemetry/Sinks/OpenTelemetry/Convert.cs
@@ -56,6 +56,7 @@ internal static class Convert
         ProcessMessage(logRecord, renderedMessage);
         ProcessMessageTemplate(logRecord, logEvent);
         ProcessLevel(logRecord, logEvent);
+        ProcessException(logRecord, logEvent);
 
         return logRecord;
     }
@@ -129,4 +130,24 @@ internal static class Convert
         logRecord.TimeUnixNano = ConvertUtils.ToUnixNano(logEvent.Timestamp);
     }
 
+    internal static void ProcessException(LogRecord logRecord, LogEvent logEvent)
+    {
+        var ex = logEvent.Exception;
+        if (ex != null)
+        {
+            var attrs = logRecord.Attributes;
+
+            attrs.Add(ConvertUtils.NewStringAttribute(TraceSemanticConventions.AttributeExceptionType, ex.GetType().ToString()));
+
+            if (ex.Message != "")
+            {
+                attrs.Add(ConvertUtils.NewStringAttribute(TraceSemanticConventions.AttributeExceptionMessage, ex.Message));
+            }
+
+            if (ex.StackTrace != null && ex.StackTrace != "")
+            {
+                attrs.Add(ConvertUtils.NewStringAttribute(TraceSemanticConventions.AttributeExceptionStacktrace, ex.ToString()));
+            }
+        }
+    }
 }
